chore(all): move long tests to integration tests (#2887)
- Move test files with >75% long tests to integration-tagged-and-named test files
- Add package names to the integration tests CI workflow matrix
- Remove usage of `t.Short()` + skip `Test_Gossip` since it doesn't work
- Move `//go:generate` commands to non-integration-tagged test files to avoid passing `-tags integration` to `go generate`
- Remove older `// +build integration`
- Remove average coverage targets in codecov configuration
