+++ .github/workflows/build.yml
@@ -102,7 +102,7 @@ jobs:
 
       - name: Generate coverage report
         run: |
-          go test ./... -short -coverprofile=coverage.out -covermode=atomic -timeout=20m
+          go test ./... -coverprofile=coverage.out -covermode=atomic -timeout=20m
       - uses: codecov/codecov-action@v3.1.1
         with:
           files: ./coverage.out
