+++ .github/workflows/integration-tests.yml
@@ -25,9 +25,14 @@ jobs:
           [
             github.com/ChainSafe/gossamer/dot,
             github.com/ChainSafe/gossamer/dot/core,
+            github.com/ChainSafe/gossamer/dot/rpc,
             github.com/ChainSafe/gossamer/dot/rpc/modules,
-            github.com/ChainSafe/gossamer/lib/babe,
+            github.com/ChainSafe/gossamer/dot/rpc/subscription,
+            github.com/ChainSafe/gossamer/dot/state,
+            github.com/ChainSafe/gossamer/dot/digest,
+            github.com/ChainSafe/gossamer/dot/network,
             github.com/ChainSafe/gossamer/dot/sync,
+            github.com/ChainSafe/gossamer/lib/babe,
             github.com/ChainSafe/gossamer/lib/grandpa,
           ]
     runs-on: ubuntu-latest
