+++ .github/workflows/unit-tests.yml
@@ -73,7 +73,7 @@ jobs:
           cd ..
 
       - name: Run unit tests
-        run: go test -short -coverprofile=coverage.out -covermode=atomic -timeout=45m ./...
+        run: go test -coverprofile=coverage.out -covermode=atomic -timeout=45m ./...
 
       - name: Test State - Race
         run: make test-state-race
