+++ dot/digest/digest_integration_test.go
@@ -1,3 +1,5 @@
+//go:build integration
+
 // Copyright 2021 ChainSafe Systems (ON)
 // SPDX-License-Identifier: LGPL-3.0-only
 
@@ -21,9 +23,6 @@ import (
 
 	"github.com/stretchr/testify/require"
 )
-
-//go:generate mockgen -destination=mock_telemetry_test.go -package $GOPACKAGE github.com/ChainSafe/gossamer/dot/telemetry Client
-//go:generate mockgen -destination=mock_grandpa_test.go -package $GOPACKAGE . GrandpaState
 
 func newTestHandler(t *testing.T) (*Handler, *state.Service) {
 	testDatadirPath := t.TempDir()
