+++ dot/network/connmgr_integration_test.go
@@ -1,3 +1,5 @@
+//go:build integration
+
 // Copyright 2021 ChainSafe Systems (ON)
 // SPDX-License-Identifier: LGPL-3.0-only
 
@@ -126,10 +128,6 @@ func TestProtectUnprotectPeer(t *testing.T) {
 }
 
 func TestPersistentPeers(t *testing.T) {
-	if testing.Short() {
-		t.Skip() // this sometimes fails on CI
-	}
-
 	t.Parallel()
 
 	configA := &Config{
@@ -165,10 +163,6 @@ func TestPersistentPeers(t *testing.T) {
 }
 
 func TestRemovePeer(t *testing.T) {
-	if testing.Short() {
-		t.Skip() // this sometimes fails on CI
-	}
-
 	t.Parallel()
 
 	configA := &Config{
@@ -204,10 +198,6 @@ func TestRemovePeer(t *testing.T) {
 }
 
 func TestSetReservedPeer(t *testing.T) {
-	if testing.Short() {
-		t.Skip() // this sometimes fails on CI
-	}
-
 	t.Parallel()
 
 	nodes := make([]*Service, 3)
