+++ dot/node_integration_test.go
@@ -9,12 +9,20 @@ import (
 	"context"
 	"encoding/hex"
 	"encoding/json"
+	"fmt"
 	"os"
 	"path/filepath"
 	"testing"
 
 	"github.com/ChainSafe/gossamer/dot/core"
+	digest "github.com/ChainSafe/gossamer/dot/digest"
+	network "github.com/ChainSafe/gossamer/dot/network"
 	"github.com/ChainSafe/gossamer/dot/state"
+	dotsync "github.com/ChainSafe/gossamer/dot/sync"
+	system "github.com/ChainSafe/gossamer/dot/system"
+	"github.com/ChainSafe/gossamer/dot/telemetry"
+	"github.com/ChainSafe/gossamer/dot/types"
+	"github.com/ChainSafe/gossamer/internal/log"
 	"github.com/ChainSafe/gossamer/lib/babe"
 	"github.com/ChainSafe/gossamer/lib/common"
 	"github.com/ChainSafe/gossamer/lib/crypto/sr25519"
@@ -22,10 +30,175 @@ import (
 	"github.com/ChainSafe/gossamer/lib/grandpa"
 	"github.com/ChainSafe/gossamer/lib/keystore"
 	"github.com/ChainSafe/gossamer/lib/runtime"
+	"github.com/ChainSafe/gossamer/lib/runtime/wasmer"
 	"github.com/ChainSafe/gossamer/lib/trie"
 	"github.com/ChainSafe/gossamer/lib/utils"
+	gomock "github.com/golang/mock/gomock"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
+
+func TestNewNode(t *testing.T) {
+	ctrl := gomock.NewController(t)
+
+	mockTelemetryClient := NewMockClient(ctrl)
+	mockTelemetryClient.EXPECT().SendMessage(gomock.Any())
+
+	initConfig := NewTestConfig(t)
+
+	genFile := NewTestGenesisRawFile(t, initConfig)
+
+	networkConfig := &network.Config{
+		BasePath:    t.TempDir(),
+		NoBootstrap: true,
+		NoMDNS:      true,
+	}
+	setConfigTestDefaults(t, networkConfig)
+
+	testNetworkService, err := network.NewService(networkConfig)
+	require.NoError(t, err)
+
+	config := state.Config{
+		Path:     initConfig.Global.BasePath,
+		LogLevel: initConfig.Log.StateLvl,
+	}
+
+	dotConfig := &Config{
+		Global:  GlobalConfig{BasePath: initConfig.Global.BasePath},
+		Init:    InitConfig{Genesis: genFile},
+		Account: AccountConfig{Key: "alice"},
+		Core: CoreConfig{
+			Roles:           common.FullNodeRole,
+			WasmInterpreter: wasmer.Name,
+		},
+	}
+
+	dotConfig.Init = InitConfig{Genesis: genFile}
+	dotConfig.Account = AccountConfig{Key: "alice"}
+	dotConfig.Core.Roles = common.FullNodeRole
+	dotConfig.Core.WasmInterpreter = wasmer.Name
+	dotConfig.Global.Name = "TestNode"
+
+	ks, err := initKeystore(t, dotConfig)
+	assert.NoError(t, err)
+
+	mockServiceRegistry := NewMockServiceRegisterer(ctrl)
+	mockServiceRegistry.EXPECT().RegisterService(gomock.Any()).Times(8)
+
+	m := NewMocknodeBuilderIface(ctrl)
+	m.EXPECT().isNodeInitialised(dotConfig.Global.BasePath).Return(nil)
+	m.EXPECT().createStateService(dotConfig).DoAndReturn(func(cfg *Config) (*state.Service, error) {
+		stateSrvc := state.NewService(config)
+		// create genesis from configuration file
+		gen, err := genesis.NewGenesisFromJSONRaw(cfg.Init.Genesis)
+		if err != nil {
+			return nil, fmt.Errorf("failed to load genesis from file: %w", err)
+		}
+		// create trie from genesis
+		trie, err := wasmer.NewTrieFromGenesis(*gen)
+		if err != nil {
+			return nil, fmt.Errorf("failed to create trie from genesis: %w", err)
+		}
+		// create genesis block from trie
+		header, err := trie.GenesisBlock()
+		if err != nil {
+			return nil, fmt.Errorf("failed to create genesis block from trie: %w", err)
+		}
+		stateSrvc.Telemetry = mockTelemetryClient
+		err = stateSrvc.Initialise(gen, &header, &trie)
+		if err != nil {
+			return nil, fmt.Errorf("failed to initialise state service: %s", err)
+		}
+
+		err = stateSrvc.SetupBase()
+		if err != nil {
+			return nil, fmt.Errorf("cannot setup base: %w", err)
+		}
+		return stateSrvc, nil
+	})
+
+	m.EXPECT().createRuntimeStorage(gomock.AssignableToTypeOf(&state.Service{})).Return(&runtime.
+		NodeStorage{}, nil)
+	m.EXPECT().loadRuntime(dotConfig, &runtime.NodeStorage{}, gomock.AssignableToTypeOf(&state.Service{}),
+		ks, gomock.AssignableToTypeOf(&network.Service{})).Return(nil)
+	m.EXPECT().createBlockVerifier(gomock.AssignableToTypeOf(&state.Service{})).
+		Return(&babe.VerificationManager{})
+	m.EXPECT().createDigestHandler(log.Critical, gomock.AssignableToTypeOf(&state.Service{})).
+		Return(&digest.Handler{}, nil)
+	m.EXPECT().createCoreService(dotConfig, ks, gomock.AssignableToTypeOf(&state.Service{}),
+		gomock.AssignableToTypeOf(&network.Service{}), &digest.Handler{}).
+		Return(&core.Service{}, nil)
+	m.EXPECT().createGRANDPAService(dotConfig, gomock.AssignableToTypeOf(&state.Service{}),
+		ks.Gran, gomock.AssignableToTypeOf(&network.Service{}),
+		gomock.AssignableToTypeOf(&telemetry.Mailer{})).
+		Return(&grandpa.Service{}, nil)
+	m.EXPECT().newSyncService(dotConfig, gomock.AssignableToTypeOf(&state.Service{}), &grandpa.Service{},
+		&babe.VerificationManager{}, &core.Service{}, gomock.AssignableToTypeOf(&network.Service{}),
+		gomock.AssignableToTypeOf(&telemetry.Mailer{})).
+		Return(&dotsync.Service{}, nil)
+	m.EXPECT().createBABEService(dotConfig, gomock.AssignableToTypeOf(&state.Service{}), ks.Babe,
+		&core.Service{}, gomock.AssignableToTypeOf(&telemetry.Mailer{})).
+		Return(&babe.Service{}, nil)
+	m.EXPECT().createSystemService(&dotConfig.System, gomock.AssignableToTypeOf(&state.Service{})).
+		DoAndReturn(func(cfg *types.SystemInfo, stateSrvc *state.Service) (*system.Service, error) {
+			gd, err := stateSrvc.Base.LoadGenesisData()
+			systemService := system.NewService(cfg, gd)
+			return systemService, err
+		})
+	m.EXPECT().createNetworkService(dotConfig, gomock.AssignableToTypeOf(&state.Service{}),
+		gomock.AssignableToTypeOf(&telemetry.Mailer{})).Return(testNetworkService, nil)
+
+	got, err := newNode(dotConfig, ks, m, mockServiceRegistry)
+	assert.NoError(t, err)
+
+	expected := &Node{
+		Name: "TestNode",
+	}
+
+	assert.Equal(t, expected.Name, got.Name)
+}
+
+func Test_nodeBuilder_loadRuntime(t *testing.T) {
+	cfg := NewTestConfig(t)
+	type args struct {
+		cfg *Config
+		ns  *runtime.NodeStorage
+		ks  *keystore.GlobalKeystore
+		net *network.Service
+	}
+	tests := []struct {
+		name string
+		args args
+		err  error
+	}{
+		{
+			name: "base case",
+			args: args{
+				cfg: cfg,
+				ns:  &runtime.NodeStorage{},
+				ks:  nil,
+				net: nil,
+			},
+			err: nil,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			ctrl := gomock.NewController(t)
+			stateSrvc := newStateService(t, ctrl)
+			no := nodeBuilder{}
+			err := no.loadRuntime(tt.args.cfg, tt.args.ns, stateSrvc, tt.args.ks, tt.args.net)
+			assert.ErrorIs(t, err, tt.err)
+			blocks := stateSrvc.Block.GetNonFinalisedBlocks()
+			for i := range blocks {
+				hash := &blocks[i]
+				code, err := stateSrvc.Storage.GetStorageByBlockHash(hash, []byte(":code"))
+				require.NoError(t, err)
+				require.NotEmpty(t, code)
+			}
+		})
+	}
+}
 
 func TestInitNode_Integration(t *testing.T) {
 	cfg := NewTestConfig(t)
