+++ dot/rpc/subscription/websocket_integration_test.go
@@ -1,6 +1,8 @@
 // Copyright 2021 ChainSafe Systems (ON)
 // SPDX-License-Identifier: LGPL-3.0-only
 
+//go:build integration
+
 package subscription
 
 import (
