+++ dot/state/storage_notify_test.go
@@ -130,10 +130,6 @@ func TestStorageState_RegisterStorageObserver_Multi_Filter(t *testing.T) {
 }
 
 func Test_Example(t *testing.T) {
-	if testing.Short() {
-		t.Skip("skipping subscription example")
-	}
-
 	// this is a working example of how to use db.Subscribe taken from
 	// https://github.com/dgraph-io/badger/blob/f50343ff404d8198df6dc83755ec2eab863d5ff2/db_test.go#L1939-L1948
 	prefix := []byte{'a'}
@@ -179,7 +175,6 @@ func Test_Example(t *testing.T) {
 	if err != nil {
 		log.Fatal(err)
 	}
-	time.Sleep(time.Second)
 	log.Printf("stopping subscription")
 	cancel()
 	log.Printf("waiting for subscription to close")
