+++ dot/sync/chain_processor_integration_test.go
@@ -1,5 +1,4 @@
 //go:build integration
-// +build integration
 
 // Copyright 2021 ChainSafe Systems (ON)
 // SPDX-License-Identifier: LGPL-3.0-only
